feat: integrate account-web to utils/account.ts
