--- account-web/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@planetarium/account-web",
-  "version": "0.0.38",
-  "license": "LGPL",
-  "source": "index.ts",
-  "module": "dist/index.mjs",
-  "main": "dist/index.cjs",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist/**/*"
-  ],
-  "imports": {
-    "#crypto": {
-      "node": "./webcrypto/node.ts",
-      "default": "./webcrypto/browser.ts"
-    },
-    "#buffer": {
-      "node": "node:buffer",
-      "default": "buffer/"
-    }
-  },
-  "dependencies": {
-    "@noble/secp256k1": "^1.7.0",
-    "buffer": "^6.0.3",
-    "ethers": "^5"
-  },
-  "devDependencies": {
-    "@planetarium/sign": "workspace:^",
-    "@types/node": "^18.7.23",
-    "nanobundle": "^0.0.28"
-  },
-  "exports": {
-    "require": "./dist/index.cjs",
-    "node": "./dist/index.mjs",
-    "import": "./dist/index.esm.js",
-    "types": [
-      "./dist/index.d.ts",
-      "./dist/util.d.ts",
-      "./dist/v3.d.ts"
-    ]
-  },
-  "peerDependencies": {
-    "@planetarium/sign": "workspace:^"
-  },
-  "scripts": {
-    "build": "nanobundle build"
-  }
-}
