--- account-web/webcrypto/browser.ts
@@ -1,1 +1,0 @@
-export const crypto = globalThis.crypto;
