--- account-web/webcrypto/node.ts
@@ -1,2 +1,0 @@
-import { webcrypto } from "node:crypto";
-export const crypto = webcrypto;
